refactor(engine): `query_spec`, so a walk over query results stops hand-rolling the skip
`list` is a candidate set: a provider may advertise an addr `get` declines.
Five call sites each re-derived that — the same `TargetNotFoundError` downcast
plus a paragraph explaining it — and the one that got it wrong is what broke
`heph i labels`.

`Engine::query_spec` is `query` one tier up: the spec of every matching
target, unresolvable candidates already dropped. `labels` is now a fold over
it and carries no error handling of its own.

The def-tier walks (`codegen_copy_overlaps`, the gitignore enumeration,
`validate`'s link check) and `revdeps` can't take a stream — they fan out
through `join_all_failable` to report *every* failure, where a stream stops at
the first — so they share the decision instead: `query::skip_unresolvable`,
which holds the one copy of the rationale, including why only a NotFound
naming *this* addr counts.

The query resolver's own speculative arm keeps its wider skip (any NotFound,
plus cycles): a candidate is being inspected only to evaluate a matcher there,
so a missing dep means "cannot match", not "broken graph".
